Enabled the moving of start and finish nodes
Functionality is handled in the handleMouseUp/Down/Enter functions. Added new states "previousNode: {row, col}, movingStart and movingFinish".

The movingStart and movingFinish are booleans that indicate if we are currently moving the start/finish nodes.

The previousNode state indicates the previous node we were on in the last iteration. This is used to update the isStart or isFinish property from the previous node to the current node.